pyproject.toml
[tool.poetry]
name = "boris-bot"
version = "0.1.0"
description = ""
authors = ["nicolatoscan <nicolatoscan99@outlook.it>"]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.20.2"
tqdm = "^4.60.0"
Pillow = "^8.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
